started adding TypeInfo check in event_call_action

+++ include/ace-vm/StackMemory.hpp
@@ -11,7 +11,7 @@ namespace vm {
 
 class Stack {
 public:
-    static const uint16_t STACK_SIZE;
+    static const size_t STACK_SIZE;
 
     friend std::ostream &operator<<(std::ostream &os, const Stack &stack);
